Select entire URL when share-url text box is clicked/focused.
The most common action when focusing the share-url text box is to 1)
select all text, 2) copy or cut it. This automates step 1 by selecting
all text whenever the text box is focused so that the user can simply
copy the url right away. I think this leads to a better user experience.