playground: Add missing supported packages.
In commit ef99944d61236035b23b1c3f1941de267f344ebf, I've run update.sh to generate the latest playground. In the diff, it's visible that 3 files were removed:

-	playground/pkg/encoding/gob.a.js
-	playground/pkg/go/build.a.js
-	playground/pkg/go/doc.a.js

That means they were previously built but not anymore with the latest version (possibly because the imports graph of standard library has changed in Go 1.5, so they are no longer dependencies, or possibly they were included erroneously). They are not explicitly listed in the list of packages to install and include.

According to the current compatibility table, 2 of those 3 packages are supported (encoding/gob and go/doc). Add them to the update.sh script so they are included.

Also add go/ast explicitly since it's supported. It was already built (as a dependency), but this will ensure it'll always be included even if it's no longer a dependency.