playground/internal/imports: add syscall/js identifiers
GopherJS 1.12-2 has added support for syscall/js package,
so it should be available as part of imports.

The syscall/js API is not included in the GOROOT/api/go1.*.txt files
at this time, and so it needs to be added to mkstdlib.go explicitly.
Use the mkapi.go program to determine the syscall/js API.

Regenerate zstdlib.go with the updated generator, using Go 1.12.13:

	go1.12.13 generate github.com/gopherjs/gopherjs.github.io/playground/internal/imports

Updates golang/go#27590